+++ main/managers/settings_manager.c
@@ -805,6 +805,10 @@ void settings_persist_setting(SettingsType setting) {
 #endif
         case SETTING_RUN_SETUP_WIZARD:
         case SETTING_I2C_SCAN:
+        case SETTING_WIGLE_TEST_API:
+        case SETTING_WIGLE_HELP:
+        case SETTING_WIGLE_MANUAL_UPLOAD:
+        case SETTING_WIGLE_STATS:
              // Actions, not saved
              return;
         case SETTING_SETUP_COMPLETE: